AP_ExternalAHRS: added frontend/backend split
allow for multiple backends
